+++ src/yarn3.cc
@@ -61,11 +61,11 @@ namespace trng {
   }
 
   const yarn3::parameter_type
-  yarn3::LEcuyer1=parameter_type(2021422057l, 1826992351l, 1977753457l);
+  yarn3::LEcuyer1=parameter_type(2021422057, 1826992351, 1977753457);
   const yarn3::parameter_type
-  yarn3::LEcuyer2=parameter_type(1476728729l, 0l, 1155643113l);
+  yarn3::LEcuyer2=parameter_type(1476728729, 0, 1155643113);
   const yarn3::parameter_type
-  yarn3::LEcuyer3=parameter_type(65338l, 0l, 64636l);
+  yarn3::LEcuyer3=parameter_type(65338, 0, 64636);
   
   // Random number engine concept
   yarn3::yarn3(yarn3::parameter_type P) :
@@ -81,7 +81,7 @@ namespace trng {
   }
 
   void yarn3::seed(unsigned long s) {
-    long long t=s;
+    int64_t t=s;
     t%=modulus;
     if (t<0)
       t+=modulus;
